`Update deploy.json and deploy.parameters.json configurations`
Refactored the `Microsoft.Web/sites` resource type in `deploy.json` by downgrading the `apiVersion` from `2020-06-01` to `2018-11-01`, adding the `kind` property with a value of `app`, and modifying the `serverFarmId` property to use the `webAppName` parameter. Removed the `hostingPlanName` and `sku` parameters along with their metadata and default values from `deploy.json`. Also, eliminated the `Microsoft.Web/serverfarms` resource type and its dependencies from `deploy.json`. In `deploy.parameters.json`, updated the value for `webAppName` to `crud-application-123` and removed the `hostingPlanName` and `sku` parameters.