Remove master_crud-application-123.yml, disabling ASP.Net Core app deployment
The YAML file `master_crud-application-123.yml` responsible for the build and deployment of an ASP.Net Core app to Azure Web App named `crud-application-123` has been removed. This file contained a GitHub Actions workflow that was triggered on push to the master branch or manually via workflow dispatch. The workflow consisted of two jobs: `build` and `deploy`. The `build` job built the application and published it as an artifact named `.net-app`. The `deploy` job then deployed this artifact to the Azure Web App in the `Production` slot. With the removal of this file, this workflow will no longer run.