Updated .NET version and improved formatting in Azure steps
Updated the `dotnet-version` to '8.x' and enabled `include-prerelease` in the `setup-dotnet` action. Added a step to restore NuGet packages using `dotnet restore`. Corrected the formatting of the `permissions` field, maintaining the `id-token` permission set to write. Reformatted the Azure login step for better readability, with `client-id`, `tenant-id`, and `subscription-id` still fetched from secrets. Improved readability in the `Deploy to Azure Web App` step by moving the `package` field to a new line, keeping the package set to the current directory.